Copy C++* -> CUDA-C++* files, unmodified.
Will update CUDA-C++* files with CUDA-specific content over next few
commits. In the event that the C++* files are updated, this commit will
serve as a convenient rebase point, allowing the CUDA-C++* files to be
updated with non-CUDA-related changes. Commits following this commit, or
a future equivalent of it, will then straightforwardly apply the
relatively simple CUDA-related changes.